kuksa_apps/kanto/requirements.txt
paho-mqtt
kuksa-viss-client
ditto-client
